fix: move audio storing functions to ProviderScreen to RecordButton component
