+++ LanguageApp/app/screens/ProviderScenarioScreen.js
@@ -3,7 +3,7 @@ import { StyleSheet, View } from "react-native";
 import { KeyboardAwareScrollView } from "react-native-keyboard-aware-scroll-view";
 import firebase from "firebase/compat/app";
 import "firebase/firestore";
-import { db, storage } from "../../firebaseSetup";
+import { db } from "../../firebaseSetup";
 
 import AppButton from "../components/AppButton";
 import AppText from "../components/AppText";
@@ -16,10 +16,7 @@ function ProviderScenarioScreen({ route }) {
   // Retrieve authenticated user information
   const { user } = useContext(AuthenticatedUserContext);
 
-  // scenario data from DB brought in from previous screens
-  const [scenario, setScenario] = useState(route.params);
-  const [scenarioLanguage, setScenarioLanguage] = useState(scenario.language);
-  const [scenarioCategory, setScenarioCategory] = useState(scenario.category);
+  const scenario = route.params;
 
   // Stores the translated text
   const [cpPrompt, setCpPrompt] = useState("");
@@ -28,67 +25,6 @@ function ProviderScenarioScreen({ route }) {
   // Contains path to audio files uploaded to Storage
   const [promptAudio, setPromptAudio] = useState();
   const [answerAudio, setAnswerAudio] = useState();
-
-  /* Method to upload audio file to Storage: 
-  https://dev.to/lankinen/expo-audio-upload-recording-to-firebase-storage-and-download-it-later-25o6 
-  Parameters: uri - URI obtained from recording on device
-              type - prompt or answer - passed from button 
-  */
-  const updateStorage = async (uri, type) => {
-    try {
-      const blob = await new Promise((resolve, reject) => {
-        const xhr = new XMLHttpRequest();
-        xhr.onload = () => {
-          try {
-            resolve(xhr.response);
-          } catch (error) {
-            console.log("error:", error);
-          }
-        };
-        xhr.onerror = (e) => {
-          console.log(e);
-          reject(new TypeError("Network request failed"));
-        };
-        xhr.responseType = "blob";
-        xhr.open("GET", uri, true);
-        xhr.send(null);
-      });
-      if (blob != null) {
-        const uriParts = uri.split(".");
-        const fileType = uriParts[uriParts.length - 1];
-        const fileName =
-          type + scenario.id + route.params.language + "." + fileType;
-        const filePath = "/cp-audio/" + fileName;
-        storage
-          .ref()
-          .child(`${filePath}`)
-          .put(blob, {
-            contentType: `audio/${fileType}`,
-          })
-          .then(() => {
-            const storagePath = "gs://capstone-language-app-fa069.appspot.com";
-            if (type == "prompt") {
-              setPromptAudio(storagePath + filePath);
-            } else {
-              setAnswerAudio(storagePath + filePath);
-            }
-          })
-          .catch((e) => console.log("error:", e));
-      } else {
-        console.log("error with blob");
-      }
-    } catch (error) {
-      console.log("error:", error);
-    }
-  };
-
-  /* Passes audio data between the screen and RecordButton component
-    adapted from https://www.kindacode.com/article/passing-data-from-a-child-component-to-the-parent-in-react/
-    Parameters: data - data to pass between the screen and component, in this case the recording uri
-                type - string prompt or answer depending which button is pressed */
-  const passLink = (data, type) => {
-    updateStorage(data, type);
-  };
 
   /* Method to update the scenario document in Firebase */
   const submitTranslation = async () => {
@@ -149,14 +85,30 @@ function ProviderScenarioScreen({ route }) {
       <KeyboardAwareScrollView>
         <View style={styles.container}>
           <ScenarioImage uriLink={scenario.image} />
+
           <AppText style={styles.text}>{scenario.prompt}</AppText>
-          <RecordButton type="prompt" passData={passLink} />
+
+          <RecordButton
+            type="prompt"
+            getReference={(reference) => setPromptAudio(reference)}
+            scenarioID={scenario.id}
+            language={scenario.language}
+          />
+
           <AppTextInput
             placeholder="Type Prompt Translation"
             onChangeText={(value) => setCpPrompt(value)}
           />
+
           <AppText style={styles.text}>{scenario.answer}</AppText>
-          <RecordButton type="answer" passData={passLink} />
+
+          <RecordButton
+            type="answer"
+            getReference={(reference) => setAnswerAudio(reference)}
+            scenarioID={scenario.id}
+            language={scenario.language}
+          />
+
           <AppTextInput
             placeholder="Type Answer Translation"
             onChangeText={(value) => setCpAnswer(value)}
